Provide getVariableImage in the common AbstractLocalVariable
